REST api for hello service
