+++ backend/src/main/java/com/m01project/taskmanager/TaskManagerApplication.java
@@ -2,8 +2,9 @@ package com.m01project.taskmanager;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class})
 public class TaskManagerApplication {
 
 	public static void main(String[] args) {
