+++ backend/src/main/java/com/m01project/taskmanager/service/HelloWorldService.java
@@ -1,0 +1,23 @@
+package com.m01project.taskmanager.service;
+
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@Service
+public class HelloWorldService {
+
+    public String sayHello() {
+        return "Hello World!";
+    }
+
+    public String sayHelloTo(String name) {
+        if(name != null && !name.isBlank()) {
+            return "hello " + name.trim();
+        }
+        else {
+            return "hello nobody!";
+        }
+    }
+}
