+++ backend/src/main/resources/application.properties
@@ -1,2 +1,7 @@
 spring.application.name=task-manager
 
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
