--- backend/src/test/java/com/m01project/taskmanager/TaskManagerApplicationTests.java
@@ -1,14 +1,0 @@
-package com.m01project.taskmanager;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TaskManagerApplicationTests {
-
-	@Test
-	void helloHelloTest() {
-
-	}
-
-}
